+++ frontend/core/engine/mailer.php
@@ -216,9 +216,6 @@ class FrontendMailer
 		$replace = array('href="' . SITE_URL . '/', 'src="' . SITE_URL . '/');
 		$content = str_replace($search, $replace, $content);
 
-		// require CSSToInlineStyles
-		require_once 'external/css_to_inline_styles.php';
-
 		// create instance
 		$cssToInlineStyles = new CSSToInlineStyles();